requirements.txt
bzr==2.6.0
catkin-pkg==0.2.10
catkin-tools==0.3.1
certifi==2017.7.27.1
cffi==1.5.2
chardet==3.0.4
cheroot==5.5.0
CherryPy==10.2.1
click==6.7
coreapi==2.3.1
coreschema==0.0.4
cryptography==1.5.3
decorator==4.0.11
Django==1.11.3
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==2.1
docutils==0.12
dronekit==2.9.0
dronekit-sitl==3.2.0
empy==3.3.2
enum34==1.1.6
Flask==0.12.2
functools32==3.2.3.post2
future==0.16.0
gps==3.16
gssapi==1.2.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
ipaddress==1.0.16
ipython-genutils==0.1.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-core==4.3.0
keyboard==0.9.13
lxml==3.7.3
Markdown==2.6.9
MarkupSafe==1.0
mercurial==3.7.3
monotonic==1.2
nbformat==4.3.0
nose==1.3.7
numpy==1.11.0
openapi-codec==1.3.2
paramiko==2.1.1
pid==2.1.1
Pillow==3.2.0
plotly==2.0.5
ply==3.8
portend==1.8
psutil==5.2.0
psycopg2==2.7.3
pyasn1==0.1.9
pycparser==2.14
pycurl==7.43.0
Pygments==2.2.0
PyJWT==1.5.2
pymavlink==2.0.6
pyserial==3.3
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.11
requests==2.18.4
rosdep==0.11.4
rosdistro==0.5.0
rosinstall==0.7.7
rosinstall-generator==0.1.11
rospkg==1.0.41
simplejson==3.10.0
six==1.10.0
tempora==1.6.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vcstools==0.1.39
virtualenv==15.0.3
Werkzeug==0.12.2
whitenoise==3.3.0
wstool==0.1.12
